feat: add db-first with test

+++ src/main/kotlin/me/davidgomes/learningcqrspoc/dto/APIRequests.kt
@@ -4,4 +4,4 @@ import java.util.UUID
 
 data class NewPerson(val name: String)
 
-data class PlaceOrder(val buyerCitizenID: UUID, val itemName: String, val quantity: ULong)
+data class PlaceOrder(val buyerCitizenID: UUID, val itemName: String, val quantity: Int)
