+++ src/main/kotlin/me/davidgomes/learningcqrspoc/dto/Order.kt
@@ -2,7 +2,7 @@ package me.davidgomes.learningcqrspoc.dto
 
 import java.util.UUID
 
-data class Order(val id: UUID, val buyerCitizenID: UUID, val itemName: String, val quantity: ULong, val status: OrderStatus)
+data class Order(val id: UUID, val buyerCitizenID: UUID, val itemName: String, val quantity: Int, val status: OrderStatus)
 
 enum class OrderStatus {
     Placed, Processed
